Allow disabling class B suport
Disabling this shrinks the code by a fair bit, so this is interesting
for low-memory applications that do not need class B.

This can also disable processing of some MAC commands. They are
completely not recognized anymore, which should be ok since these
commands should never be received at all. If they are received, this
will break processing of subsequent commands, though.