Arduino: Add Arduino-specific files
This adds an implementation of the LMIC HAL, plus some files needed for
the Arduino library format and some examples.

In addition, an export.sh script is added that can be used to generate
an Arduino library from this repository. This script only reorders
files, it does not change anything. When given the --link option, it
creates symlinks instead of copies, allowing to test things in the
Arduino IDE easily, while committing the original files in this
repository.

The HAL, examples and library.properties files are taken from the
https://github.com/matthijskooijman/arduino-lmic repository, revision
90bc049 (Print downlink ack status in ttn-otaa example too). Only the
configuration mechanism was changed slightly.