Arduino: Simplify examples
This stops using the osjob scheduling from lmic for scheduling
transmissions, since this is a bit tricky to integrate with existing
code, and users now often think that using it is mandatory. By removing
it, and replacing it by simple time-based checks for scheduling
transmissions, it should be more clear how this works.

Also, remove some code specific to the Pinoccio board.