Add more verbose debug output
This is triggered by a CFG_DEBUG_VERBOSE define and gives more spammy
debug output.

This also adds some sleeping-related debug output, but this is disabled
for now since the Arduino port does not do sleeping yet.

The osxtime_t values (which are 64-bit) are printed as ostime_t (which
are 32-bit). Printing them as 64-bit would make it harder to compare
them with ostime_t values. Things might get weird or incorrect after an
overflow, though.