Let printf %F accept s4_t instead of int
In practice, u4_t and s4_t arguments are always passed to %F (freq
mostly). On systems where int is s4_t, this works fine (the signed vs
unsigned mismatch works out as long as values are smaller than 2^31).
However, on systems where int is s2_t, this causes the printf arguments
to run out of sync, producing garbage output and potentially infinite
loops.

In two cases, %F was used for a s1_t, which must now be ucast to s4_t
(it would already be implicitly converted to int, so again this was not needed
on systems that have int equal to s4_t).