Sx126x: Enable DIO3 als TCXO control
This should probably be made configurable, but for now just always
enable it (since typically DIO3 is not used as anything else anyway).

The voltage and timeout values were taken from other example code and
should probably also be configurable. Also, the timout adds TX/RX delay,
so should probably be accounted for.