Schedule RX/TX timeout after starting RX/TX
Previously, these timeouts were started before starting RX or TX. This
worked normally, but when a HAL did not have access to the BUSY pin and
instead resorted to conservative delays between SPI transactions,
starting RX or TX would take so long that the timeout would effectively
end up too short.

Note that there is no HAL committed yet that supports working without a
BUSY pin, but this came up in some as of yet uncommitted work on the
Arduino port. Even if running without a BUSY pin will never be
supported, setting the timeout like this should be fine, since in the
worst case it only effectively extends the timeout by a few ms. Since
the timeout should normally occur, this is fine.