Arduino: Support separate RX and TX enable pins
Some radio modules have separate pins for this, which allows switching
the antenna switch off entirely to save power. On boards with just a
single RXTX pin, this pin can be set as TX pin, to fall back to RX mode
when idle.

This commit intentionally breaks compatibility with the pinmap, to
force sketches to update their pinmap (also for the busy pin, since
omitting a field at the end of a struct initializes it to 0, which could
be a valid pin number).