Arduino: Add pinout for some standard boards
These are boards that include a LoRa module, so have a well-known
pinout.

These are added to a separate file, which is copied into each example
directory on export, which makes it easier to change these pinmaps in a
single place.

This commit also adds some comments about how to add your own pinmap.