Arduino: Fix NSS initialization
Previously, the NSS pin would only be set to OUTPUT mode, causing it to
become OUTPUT LOW and starting an SPI transaction already. In some cases
(possibly when the chip was subsequently reset), the first SPI
transaction would then fail (readReg would return 0), causing
an assertion failure during initialization.

This is fixed by writing the NSS pin HIGH in addition to changing to
OUTPUT mode. To prevent a low pulse on the line, it is written HIGH
before setting OUTPUT, but it seems not all cores support this (e.g. the
arduino-STM32L4 reset to LOW when changing to OUTPUT), so write HIGH
again after setting OUTPUT to be sure.