Fix undefined casting in GetRandom
In C, it is not allowed to reference variables using pointers to an
unrelated type, except when the pointer is uint8_t. So dereferencing a
uint8_t[] through uint32_t* is not allowed (and gives a warning about
type-punning), but dereferencing uint32_t through a uint8_t* is allowed.

This rewrites GetRandom slightly to fix this warning.