Let debug_printf accept long modifier (e.g. %ld)
This allows printing long values. On ARM, these are identical to int,
but on AVR they are different, so this allows printing 32-bit values
there.

This is not actually used anywhere yet (the only 32-bit values used are
timestamps, which are printed using %t), but this prepares for printing
raw timestamps in a subsequent commit.