Sx126x: Fix compilation failure due to radio_cw rename
In version 2.2, Semtech changed the 127x driver to rename txsw to
radio_cw, but this change was not made in the 126x driver, making
compilation fail when using it.

This renames the function for 126x to at least fix compilation.

However, for 126x the continuous mode was split into a TXCW and TXCONT
mode, where the former uses the existing radio_cw (formerly txcw)
function, but the latter txfsk() or txlora() with some changes for
continuous mode. This is not implemented for 126x, so TXCONT is probably
broken for 126x now.