Arduino: Rename top-level include to basicmac.h
This is the first header from this library that an Arduino sketch should
include, allowing the IDE to detect the sketch needs this library.
Renaming this from the lmic.h that LMIC uses, allows LMIC and BasicMAC
to co-exist in your libraries folder, with sketches selecting the right
library based on this include.

Note that other header files still use 'lmic' in their name (and
contents). Renaming all these would be too invasive, but also not needed
(once the top-level include selected one library, all other includes
will be resolved using that library if possible).

This is a breaking change, which breaks any existing sketches. Fixing
such sketches is a matter of replacing `#include <lmic.h>` with
`#include <basicmac.h>`.