Arduino: Rename examples to basicmac-abp/otaa
These examples were not really TTN-specific, but could just as easily be
used for other networks (maybe needing some settings changes), so
reflect that in the name. This adds "basicmac" to the name, to make it
easier to distinguish them from the lmic examples too, when working with
both libraries.