Make rx1DrOff signed
This array stores offsets between up and own datarates. Some of these
values are negative, but were now stored as unsigned numbers. In
practice, this worked out ok because unsigned and signed addition are
the same, and the result of such addition (in prepareDnDr) was cast to
s1_t (so any upper bits incorrect due to lack of sign extension were
discarded anyway).

However, in a few cases, comparisons with ILLEGAL_RX1DRoff (which is
defined as -128) would *not* work. The compiler also warned about this:

    warning: comparison is always false due to limited range of data type [-Wtype-limits]
	 if( REGION.rx1DrOff[LMIC.dn1DrOffIdx] == ILLEGAL_RX1DRoff )
                                           ^~

And then the optimizer would remove the if, assuming it is always false.

In practice, this only meant that any invalid offsets in a join accept packet
or MCMD_DN2P_SET MAC command would not be rejected.