Fix potential infinite loop in rps2dndr
If this function would be passed rps that matches none of the defined
DR's, this loop would loop infinitely. This happened because dr_t is
signed, so `dr--` would loop from 0 to 255 rather than to -1, and `dr >=
0` would never be false.

The compiler warned about this:

    warning: comparison is always true due to limited range of data type [-Wtype-limits]
	 for( dr_t dr = 15; dr >= 0; dr-- ) {
			       ^~

In practice, this never happened because there would always be a
matching dr to break the loop, but fix this anyway.