Make RPS-building more clear
This renames the LWUPDR/LWDNDR macros to LORA_UP_RPS and LORA_DN_RPS to
make it more clear that these parameters are LoRa-specific and build an
RPS rather than an DR-number. MAKERPS is also renamed to MAKE_LORA_RPS
and makeRps to makeLoraRps.

For FSK, FSK_UP_RPS, MAKE_FSK_RPS and makeFskRps are introduced which
accept only the parameters needed for FSK.