Calculate LMIC.dndr earlier
Previously, LMIC.dndr was set to the TX datarate on TX, and then the
corresponding RX1 datarate was derived in txDone. Since nothing happens
in between that could influence the RX1 datarate, it is easier to just
derive the RX1 datarate on TX already.

This removes the confusion where `LMIC.dndr` would actually contain the
TX datarate. It also prepares for introducing a custom datarate.