debug: Handle a zero framecounter properly
The `LMIC.seqnoUp` is the frame counter to be used for the *next* uplink
packet. Therefore, one is subtracted before printing, to get the counter
of the current framecounter.

However, when no data uplinks have been sent yet (e.g. when sending a
join request), `LMIC.seqnoUp` would be 0 and the subtraction would
underflow. Earlier, this would print -1, but now that the counter is
properly printed as unsigned, this would print some large value.

With this commit, simply print 0 when no frames have been sent yet.