sx126x: Fix random write in GetRandom
In commit f74ec70 (Fix undefined casting in GetRandom), an addressof
operator for `value` was forgotten, so instead of casting the pointer to
`value`, this would cast `value` itself to a pointer. Combined with a
lack of initialization on `value`, this would end up writing the random
value to a random place in memory, and not actually returning it.

While we're here, also use sizeof instead of hardcoding the number of
bytes to read.

Note that none of this actually seems to make GetRandom work, the
hardware still returns only zeroes from the random register somehow...