sx126x: No longer use DIO3 as TCXO control by default
Enabling this unconditionally originally seemed harmless, since the pin
was otherwise unused, but this also switches the transceiver into TCXO
mode, breaking operation with a crystal.

Instead of enabling this unconditionally, it is now disabled by default
and can be enabled explicitly. On Arduino, set the `tcxo` field of the
pin map to `LMIC_CONTROLLED_BY_DIO3`. With Makefile-based stm32, define
`LMIC_DIO3_CONTROLS_TCXO` on the compiler commandline.

The API between the radio driver and the HAL is not ideal and might need
to be reconsidered later, but since the change on the Arduino side is
breaking, better to get that out of the way soon (and the HAL side can
be improved later).

This fixes #9