hal: No longer use DIO2 to control RXTX by default
Enabling this unconditionally originally seemed harmless, since the pin was
otherwise unused, but there is a (tiny) chance that boards have unused DIO pins
grounded, which could lead to short-circuits. Also, it is probably good to make
the DIO2-controls-TXRX situation explicit in the pinmap and be consistent with
DIO3-controls-TCXO.

Instead of enabling this unconditionally, it is now disabled by default
and can be enabled explicitly. On Arduino, set the `tx` field of the
pin map to `LMIC_CONTROLLED_BY_DIO2`. With Makefile-based stm32, define
`LMIC_DIO2_CONTROLS_RXTX` on the compiler commandline.

The API between the radio driver and the HAL is not ideal and might need
to be reconsidered later, but since the change on the Arduino side is
breaking, better to get that out of the way soon (and the HAL side can
be improved later).