arduino: Show how to select SX1261
Previously, target-config only showed the SX1262, not the SX1261. This
fixes that by adding a commented define. It also completes the guard
around the entire block, which is relevant for automatic (Github Action)
builds (before this change both `BRD_sx1261_radio` and
`BRD_sx1262_radio` might end up defined, which did not cause a
compilation failure, so it was not noticed before).