arduino: Change license to BSD
Some parts of this code were still marked with the EPL license, used by
older versions of LMIC. This changes the license to 3-clause BSD, the
same as the rest of LMIC. This license is also added to the files that
previously did not state any license.

All of the affected code was written by me, with no significant
contributions from others.