sx126x: Use random generator for random seed
This radio has a random generator built in, so use it to seed the
internal random generator. For the sx127x, the old, imperfect, method (based on
DevEUI and compilation time is still used. This could maybe use RSSI
measurements like LMIC did?

Previously, the random generator was seeded on first use, but to prevent
needing the radio while it is already operating, this now seeds the
generator on startup.

In particular, this means that on sx126x, the channel selection and join
nonces are now properly randomized.