sx126x: Fix GetRandom
When using DIO2 or DIO3 to control the TCXO or RxTx switch and GetRandom
is called while the radio sleeps, the random generator will not be able
to receive noise and return only zeroes.

This fixes this by setting up the settings whenever GetRandom is called.

To make sure CommonSetup is visible, also move GetRandom a bit downward
in the file.