Make `ostime_t` and `osxtime_t` unsigned
In C, overflow (or underflow) of signed integers is undefined. In
practice, due to the 2s complement that is (now) used everywhere, signed
integers will wrap just like unsigned integers, but because it is
undefined compilers can optimize such overflows to return any value at
all. For this reason, signed integers should not be manipulated so that
they can overflow.

For this reason, `ostime_t` and `osxtime_t` should really be unsigned,
so that their manipulations have defined overflow behavior.

When subtracting two timestamps when their relative order is
unknown (i.e. either one might be earlier than the other), the overflow
behavior is now well-defined, but the result should typically be
converted to a signed value to detect whether one or the other was
earlier (i.e. the result is positive or negative). Previously, this was
implicit, because subtracting two signed values produced a signed
result, but now this must be done explicitly by casting the result of
the subtraction (not the arguments, since the subtraction must remain
unsigend).

Most of this commit consists of adding these casts.

Note that for other manipulations, such as adding or subtracting an
*interval* to a timestamp, or subtracting an earlier timestamp from a
later timestamp, the unsigned overflow will produce the right result
right away without casts, so most uses of timestamps can remain
unmodified.

Also, for `osxtime_t`, timestamps are usually directly compared (rather
than subtracted), which works because these timestamps can be assumed to
never overflow. Such comparisons also remain valid with unsigned
timestamps, so no changes needed there either.

For this commit, I checked *all* uses of variables of the `ostime_t` or
`osxtime_t` types to identify cases where additional casts were needed.

I also checked that all calls to the conversions functions (e.g.
us2osticks) use unsigned or have necessarily positive arguments, so no
signed versions of those are needed.

Finally, I looked at the compiled code, which has a lot of changes in
code locations and register allocations, but the only actual change I
could identify was that some sign extension was replaced by zero
extension, as can be expected.